Remove DynamicSupervisor and turn into library
VintageNetDirect was adding `OneDHCPD.Server` to its supervision tree
rather than using the `start_server` method. This was the intention, but
some old code was still present that made OneDHCPD an OTP Application.
This removes that code. OneDHCPD is now a regular library and if it's
not used, it won't be starting a DynamicSupervisor any more.